Update performance/security docs, fix build issues, add benchmarks
- PERFORMANCE.md: Add Section 9.0 with Debian 13 benchmark results
  - Collective operations: Broadcast 10.73 GB/s, Reduce 10.41 GB/s peak
  - Stress test: 4,591 ops/sec, 9.63 GB/s throughput
- SECURITY.md: Add Appendix D with automated security scan results
  - No critical vulnerabilities (gets=0, system=0)
  - 162 snprintf vs 31 sprintf (good safe-practice ratio)
- src/device/Makefile: Add gensrc include path fix
- src/include/perf_counters.h: Fix C++ header organization
- src/transport/net.cc: Fix ssize_t type comparison warning
- tests/benchmark/: Add cross-platform NCCL stress test suite
- tests/platform/: Add Linux platform test binaries (Debian 13)

+++ src/transport/net.cc
@@ -1791,7 +1791,7 @@ static ncclResult_t recvProxyProgress(struct ncclProxyState *proxyState, struct 
                   continue;
                 sub->regBufferReady = 1;
                 ptrs[subCount] = sub->recvbuff + sub->posted * NCCL_MAX_NET_SIZE;
-                sizes[subCount] = std::min(NCCL_MAX_NET_SIZE, (ssize_t)(sub->nbytes - sub->posted * NCCL_MAX_NET_SIZE));
+                sizes[subCount] = std::min((ssize_t)NCCL_MAX_NET_SIZE, (ssize_t)(sub->nbytes - sub->posted * NCCL_MAX_NET_SIZE));
               }
               else
               {
